Delete batch_norm experiment and add seq_bn_weight experiment
