+++ experiments/seq_bn_weight/job.sbatch
@@ -1,5 +1,5 @@
 #!/bin/sh
-#SBATCH --partition=gpu --gres=gpu:1 --mem=64GB --time=2-00:00:00 -o /scratch/ttrine/fish/experiments/batch_norm/log.out
+#SBATCH --partition=gpu --gres=gpu:1 --mem=64GB --time=2-00:00:00 -J seq_bn_weight -o /scratch/ttrine/fish/experiments/seq_bn_weight/log.out
 cd /scratch/ttrine/fish/
 module load python/2.7.12/b2 blas hdf5/1.8.14 cudnn/7.5 git opencv
 module unload python/2.7.10
@@ -7,4 +7,4 @@ source /scratch/ttrine/fish/venv/bin/activate
 PYTHONPATH=/scratch/ttrine/fish/venv/lib/python2.7/site-packages:$PYTHONPATH
 export PYTHONPATH
 export THEANO_FLAGS=mode=FAST_RUN,device=gpu,floatX=float32,exception_verbosity=high
-python -m experiments.batch_norm.train 10000 16 512
+python -m experiments.seq_bn_weight.train 10000 16 512
